feat: added `value` input prop
This change allows you to set a `value` that you can use to forcibly update the UI's displayed
selected item. However, this does not introduce a breaking change. This means that `onSelect`
_will_ visually update the UI, THEN the `value` update will take over and update the UI.

Keep in mind that we _will_ be changing this behavior in the next version release.
If you are not currently using `value`, I suggest updating to it, as it will be required
in the next major semver update.

Closes #1